test: agregar tests unitarios para Profile, ProfileModule y ProfileModuleItem
- Tests de integración para Profile CRUD (Update, Delete, Show)
- Tests para ProfileModule: list modules, sync modules
- Tests para ProfileModuleItem: list, sync, attach, detach items
- Seed completo de 5 tablas relacionadas
- Skip intencional: tests que requieren características de PostgreSQL

Cobertura: ~50% (lógica de negocio + integración básica con SQLite)

+++ routes/profile/module_handler.go
@@ -165,7 +165,7 @@ func (h *ModuleHandler) Sync(c *gin.Context) {
 			profileModules := make([]model.ProfileModule, len(req.Modules))
 			for i, moduleID := range req.Modules {
 				profileModules[i] = model.ProfileModule{
-					ProfileID: profileID,
+					ProfileID: uint(profileID),
 					ModuleID:  moduleID,
 				}
 			}
